+++ packages/common/internal/sparklineInteractiveBuilder.tsx
@@ -36,28 +36,43 @@ const DATE_TIME_OPTIONS = {
 
 const periodsAlt = [
   {
-    label: '1H',
+    label: '1 Std.',
     value: 'hour' as const,
   },
   {
-    label: '1D',
+    label: '1 Tag',
     value: 'day' as const,
   },
   {
-    label: '1 SEM.',
+    label: '1 Wo.',
     value: 'week' as const,
   },
   {
-    label: '1 MÊS',
+    label: '1 Mon.',
     value: 'month' as const,
   },
   {
-    label: '1 ANO',
+    label: '1 Jahr',
     value: 'year' as const,
   },
   {
     label: 'All',
     value: 'all' as const,
+  },
+];
+
+const subsetOfPeriods = [
+  {
+    label: '1D',
+    value: 'day' as const,
+  },
+  {
+    label: '1M',
+    value: 'month' as const,
+  },
+  {
+    label: '1Y',
+    value: 'year' as const,
   },
 ];
 
@@ -150,6 +165,7 @@ type SparklineInteractiveBuilderProps = {
   >;
   isMobile?: boolean;
   alternatePeriods?: boolean;
+  smallerPeriodSet?: boolean;
 };
 
 function numToLocaleString(num: number) {
@@ -162,6 +178,7 @@ export const sparklineInteractiveBuilder = ({
   SparklineInteractive,
   isMobile,
   alternatePeriods,
+  smallerPeriodSet,
 }: SparklineInteractiveBuilderProps) => {
   return memo(({ defaultPeriod, hideHoverDate, ...props }: SparklineInteractivePriceProps) => {
     // not supported onAndroid
@@ -200,10 +217,22 @@ export const sparklineInteractiveBuilder = ({
       return `$${numToLocaleString(parseInt(amount as string, 10))}`;
     }, []);
 
+    const configurablePeriods = useMemo(() => {
+      if (alternatePeriods) {
+        return periodsAlt;
+      }
+
+      if (smallerPeriodSet) {
+        return subsetOfPeriods;
+      }
+
+      return periods;
+    }, []);
+
     return (
       <SparklineInteractive
         {...props}
-        periods={alternatePeriods ? periodsAlt : periods}
+        periods={configurablePeriods}
         formatDate={formatDateWithConfig}
         formatHoverDate={!hideHoverDate ? formatHoverDate : undefined}
         defaultPeriod={defaultPeriod ?? DEFAULT_PERIOD}
@@ -253,11 +282,13 @@ export const sparklineInteractiveWithHeaderBuilder = ({
   SparklineInteractiveHeader,
   isMobile,
   alternatePeriods,
+  smallerPeriodSet,
 }: SparklineInteractiveWithHeaderBuilderProps) => {
   const SparklineInteractiveBuild = sparklineInteractiveBuilder({
     SparklineInteractive,
     isMobile,
     alternatePeriods,
+    smallerPeriodSet,
   });
 
   return memo((props: SparklineInteractivePriceProps) => {