+++ packages/common/types/LottieStatusAnimationProps.ts
@@ -1,5 +1,6 @@
 import { DimensionValue } from './DimensionStyles';
 import { NoopFn } from './Helpers';
+import { SharedProps } from './SharedProps';
 
 export type LottieStatusAnimationType =
   | 'loading'
@@ -29,4 +30,5 @@ type StatusAnimationWithHeight = {
   height: DimensionValue;
 } & BaseStatusAnimationProps;
 
-export type LottieStatusAnimationProps = StatusAnimationWithWidth | StatusAnimationWithHeight;
+export type LottieStatusAnimationProps = (StatusAnimationWithWidth | StatusAnimationWithHeight) &
+  SharedProps;