+++ packages/mobile/illustrations/Illustration.tsx
@@ -1,17 +1,21 @@
 import React, { memo, useMemo } from 'react';
-import { IllustrationBaseProps, IllustrationVariant } from '@cbhq/cds-common/types';
+import { IllustrationVariant } from '@cbhq/cds-common/types';
 import heroSquares from '@cbhq/cds-illustrations/__generated__/heroSquare/data/svgJsMap';
 import pictograms from '@cbhq/cds-illustrations/__generated__/pictogram/data/svgJsMap';
 import spotRectangles from '@cbhq/cds-illustrations/__generated__/spotRectangle/data/svgJsMap';
 import spotSquares from '@cbhq/cds-illustrations/__generated__/spotSquare/data/svgJsMap';
 
-import { createIllustration } from './createIllustration';
+import { createIllustration, IllustrationBasePropsWithA11y } from './createIllustration';
 
 const illustrations = {
   heroSquare: heroSquares,
   spotSquare: spotSquares,
   spotRectangle: spotRectangles,
   pictogram: pictograms,
+};
+
+type IllustrationProps<Type extends IllustrationVariant> = IllustrationBasePropsWithA11y<Type> & {
+  type: Type;
 };
 
 /**
@@ -22,7 +26,7 @@ const illustrations = {
 export const Illustration = memo(function Illustration<Type extends IllustrationVariant>({
   type,
   ...props
-}: { type: Type } & IllustrationBaseProps<Type>) {
+}: IllustrationProps<Type>) {
   const Component = useMemo(() => createIllustration(type, illustrations[type]), [type]);
 
   return <Component {...props} />;