+++ packages/mobile/illustrations/Pictogram.tsx
@@ -1,11 +1,13 @@
+import { PictogramProps as PictogramCommonProps } from '@cbhq/cds-common';
 import pictogramIllustrations from '@cbhq/cds-illustrations/__generated__/pictogram/data/svgJsMap';
 
-import { createIllustration } from './createIllustration';
+import { createIllustration, IllustrationA11yProps } from './createIllustration';
+
+export type PictogramProps = PictogramCommonProps & IllustrationA11yProps;
 
 /**
  * @changelog ./createIllustration.tsx, ./Illustration.tsx
  */
 export const Pictogram = createIllustration('pictogram', pictogramIllustrations);
 
-export type { PictogramProps } from '@cbhq/cds-common/types/IllustrationProps';
 export type { PictogramName } from '@cbhq/cds-illustrations';