+++ packages/ui-mobile-visreg/src/detox/isExpectedIosDevice.ts
@@ -3,7 +3,7 @@ import { DetoxConfig } from '../detoxConfig';
 // TODO: make more robust using device specs instead of just name
 export default function isExpectedIosDevice(config: DetoxConfig) {
   return (
-    config.devices.simulator.device.type === 'iPhone 12' &&
-    config.devices.simulator.type === 'ios.simulator'
+    config.devices['simulator.14'].device.type === 'iPhone 14' &&
+    config.devices['simulator.14'].type === 'ios.simulator'
   );
 }