+++ packages/web/animation/__tests__/Lottie.test.tsx
@@ -1,9 +1,13 @@
-import { render } from '@testing-library/react';
+import { render, screen } from '@testing-library/react';
 import { nux } from '@cbhq/cds-lottie-files/nux';
 
 import { Lottie } from '../Lottie';
 
 describe('Lottie', () => {
+  it('renders Lottie', () => {
+    render(<Lottie source={nux} testID="lottie-test" />);
+    expect(screen.getByTestId('lottie-test')).toBeTruthy();
+  });
   it('renders a div', () => {
     const { container } = render(<Lottie source={nux} />);
     // It's generally discouraged to use the container or DOM traversing, however this is explicitly testing that a div is created.