+++ packages/web/illustrations/Pictogram.tsx
@@ -1,11 +1,12 @@
+import type { PictogramProps as CommonPictogramProps } from '@cbhq/cds-common/types/IllustrationProps';
 import pictogramVersionMap from '@cbhq/cds-illustrations/__generated__/pictogram/data/versionMap';
 
-import { createIllustration } from './createIllustration';
+import { createIllustration, IllustrationA11yProps } from './createIllustration';
 
 /**
  * @changelog ./createIllustration.tsx, ./Illustration.tsx
  */
 export const Pictogram = createIllustration('pictogram', pictogramVersionMap);
 
-export type { PictogramProps } from '@cbhq/cds-common/types/IllustrationProps';
+export type PictogramProps = CommonPictogramProps & IllustrationA11yProps;
 export type { PictogramName } from '@cbhq/cds-illustrations';