+++ packages/common/CHANGELOG.md
@@ -2,21 +2,13 @@
 
 > [NPM registry](https://registry-npm.cbhq.net/-/web/detail/@cbhq/cds-common)
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented in this file!
 
 `@cbhq/cds-common` adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- template-start -->
 
 ## 5.0.0 ((6/5/2023, 04:33 PM PST))
-
-This is an artificial version bump with no new change.
-
-#### 📘 Misc
-
-- Rebase sparkline-mobile decomposition on release branch.
-
-## 5.0.0-rc.2 (3/23/2023, 06:36 PM PST)
 
 #### 💥 Breaking
 
@@ -27,11 +19,65 @@ This is an artificial version bump with no new change.
 
 - **[ui-mobile-playground, ui-mobile-visreg]** Fix Npm Pack. [#1841](https://github.cbhq.net/frontend/cds/pull/1841)
 
-## 5.0.0-rc.1 (3/20/2023, 10:29 AM PST)
-
 #### 💥 Breaking
 
 - **[ui-mobile-playground]** Separate UI code from visreg code for playgrounds. [#1810](https://github.cbhq.net/frontend/cds/pull/1810)
+- Enabled storybook testing for mobile for various configurations. [[#2009](https://github.cbhq.net/frontend/cds/pull/2009)] [[DX-1627](https://jira.coinbase-corp.com/browse/DX-1627)]
+
+## Unreleased
+
+#### 📘 Misc
+
+- Fix spacing comparison w/ mono-pipeline.
+
+## 4.2.14 ((6/4/2023, 12:57 PM PST))
+
+This is an artificial version bump with no new change.
+
+#### 📘 Misc
+
+- Rebase sparkline-mobile decomposition on release branch.
+
+## 4.2.13 (5/23/2023 PST)
+
+#### 🐞 Fixes
+
+- Add testID to LottieStatusAnimationProps. [[#2001](https://github.cbhq.net/frontend/cds/pull/2001)] [[DX-1071](https://jira.coinbase-corp.com/browse/DX-1071)]
+
+## 4.2.12 (5/23/2023 PST)
+
+#### 🐞 Fixes
+
+- Add testID for ListCellFallback props. [[DX-1070](https://jira.coinbase-corp.com/browse/DX-1070)]
+
+## 4.2.11 (5/18/2023, 04:33 PM PST)
+
+This is an artificial version bump with no new change.
+
+#### 📘 Misc
+
+- Updated some shared mocks to look like real data.
+
+## 4.2.10 (5/17/2023 PST)
+
+#### 🐞 Fixes
+
+- **[ModalHeader]** Remove hard coded accessibility labels. [[#1995](https://github.cbhq.net/frontend/cds/pull/1995)] [[DX-1418](https://jira.coinbase-corp.com/browse/DX-1418)]
+
+## 4.2.9 ((5/17/2023, 08:34 AM PST))
+
+This is an artificial version bump with no new change.
+
+## 4.2.8 (5/16/2023 PST)
+
+#### 🐞 Fixes
+
+- Move 'alt' prop to platform specific props. [[#1984](https://github.cbhq.net/frontend/cds/pull/1984)] [[DX-1126](https://jira.coinbase-corp.com/browse/DX-1126)]
+- Fix CardMedia generator to implicitly pass 'alt' for illustration. [[#1984](https://github.cbhq.net/frontend/cds/pull/1984)] [[DX-1126](https://jira.coinbase-corp.com/browse/DX-1126)]
+
+## 4.2.7 ((5/15/2023, 04:29 PM PST))
+
+This is an artificial version bump with no new change.
 
 ## 4.2.6 (4/20/2023 PST)