+++ packages/web/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 > [NPM registry](https://registry-npm.cbhq.net/-/web/detail/@cbhq/cds-web)
 
-All notable changes to this project will be documented in this file.
+All notable changes to this project will be documented in this file!
 
 `@cbhq/cds-web` adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
@@ -10,17 +10,9 @@ All notable changes to this project will be documented in this file.
 
 ## 5.0.0 ((6/5/2023, 04:33 PM PST))
 
-This is an artificial version bump with no new change.
-
 #### 💥 Breaking
 
 - [VStack] Updated VStack styling for when as="ul" is enabled [#1922](https://github.cbhq.net/frontend/cds/pull/1922), [DX-990](https://jira.coinbase-corp.com/browse/DX-990)
-
-#### 📘 Misc
-
-- Rebase sparkline-web decomposition on release branch.
-
-## 5.0.0-rc.2 (3/23/2023, 06:36 PM PST)
 
 #### 💥 Breaking
 
@@ -31,11 +23,67 @@ This is an artificial version bump with no new change.
 
 - **[ui-mobile-playground, ui-mobile-visreg]** Fix Npm Pack. [#1841](https://github.cbhq.net/frontend/cds/pull/1841)
 
-## 5.0.0-rc.1 (3/20/2023, 10:29 AM PST)
-
 #### 💥 Breaking
 
 - **[ui-mobile-playground]** Separate UI code from visreg code for playgrounds. [#1810](https://github.cbhq.net/frontend/cds/pull/1810)
+
+## 4.2.14 ((6/4/2023, 12:57 PM PST))
+
+This is an artificial version bump with no new change.
+
+#### 📘 Misc
+
+- Convert Divider story from mdx to tsx. [[#2007](https://github.cbhq.net/frontend/cds/pull/2007)] [[DX-192](https://jira.coinbase-corp.com/browse/DX-192)]
+
+#### 📘 Misc
+
+- Update documentation in the README. [[#2006](https://github.cbhq.net/frontend/cds/pull/2006)] [[DX-1620](https://jira.coinbase-corp.com/browse/DX-1620)]
+
+## 4.2.13 (5/23/2023 PST)
+
+#### 🐞 Fixes
+
+- Add testID to LottieStatusAnimation & test coverage. [[#2001](https://github.cbhq.net/frontend/cds/pull/2001)] [[DX-1071](https://jira.coinbase-corp.com/browse/DX-1071)]
+
+## 4.2.12 (5/23/2023 PST)
+
+#### 🐞 Fixes
+
+- Add testID for ListCellFallback & tests. [[DX-1070](https://jira.coinbase-corp.com/browse/DX-1070)]
+
+## 4.2.11 (5/18/2023, 04:33 PM PST)
+
+This is an artificial version bump with no new change.
+
+## 4.2.10 (5/17/2023 PST)
+
+#### 🐞 Fixes
+
+- **[ModalHeader]** Remove hard coded accessibility labels. [[#1995](https://github.cbhq.net/frontend/cds/pull/1995)] [[DX-1418](https://jira.coinbase-corp.com/browse/DX-1418)]
+
+## 4.2.9 (5/17/2023 PST)
+
+#### 🐞 Fixes
+
+- [trivial] fix: added dangerouslySetClassName to TableCell. [[#1991](https://github.cbhq.net/frontend/cds/pull/1991)] [[DX-1401](https://jira.coinbase-corp.com/browse/DX-1401)]
+
+## 4.2.8 (5/16/2023 PST)
+
+#### 🐞 Fixes
+
+- Fix Illustration web props to have alt prop since removed from common props. [[#1984](https://github.cbhq.net/frontend/cds/pull/1984)] [[DX-1126](https://jira.coinbase-corp.com/browse/DX-1126)]
+
+## 4.2.7 ((5/15/2023, 04:29 PM PST))
+
+This is an artificial version bump with no new change.
+
+#### 📘 Misc
+
+- Docs(Sparkline): add example for stacking sparkline. [[DX-1283](https://jira.coinbase-corp.com/browse/DX-1283)]
+
+#### 📘 Misc
+
+- **[Icon]** Split up icon story so percy will upload snapshot again. [[#1972](https://github.cbhq.net/frontend/cds/pull/1972)] [[DX-1271](https://jira.coinbase-corp.com/browse/DX-1271)]
 
 ## 4.2.6 (4/20/2023 PST)